+++ Makefile
@@ -8,6 +8,9 @@
 build:
 	go build ./...
 
+experimental:
+	go build -tags=experimental ./...
+
 linter:
 	go install github.com/golangci/golangci-lint/v2/cmd/golangci-lint@v2.0.2
