--- docs/resources/morpheus_role.md
@@ -1,65 +1,0 @@
----
-page_title: "hpe_morpheus_role Resource - terraform-provider-hpe"
-subcategory: "morpheus"
-description: |-
-  
----
-# hpe_morpheus_role (Resource)
-
-
-
-`hpe_morpheus_role` resource ....
-
-## Example Usage
-
-```terraform
-resource "hpe_morpheus_role" "example" {
-	name = "ExampleRole"
-	multitenant = false
-        description = "An example role"
-        role_type = "user"
-}
-```
-
-## Example Usage (Using legacy Morpheus Terraform provider)
-
-```terraform
-data "morpheus_task" "example_legacy_task" {
-  name = "example_task"
-}
-
-resource "hpe_morpheus_role" "example_with_legacy_provider" {
-  name = "ExampleRoleWithLegacyProvider"
-  description = "An example role using legacy provider"
-  role_type = "user"
-  permissions = jsonencode({
-    "taskPermissions" : [
-      {
-        "id" = data.morpheus_task.example_legacy_task.id
-        "access" = "full"
-      }
-    ]
-    }
-  )
-}
-```
-
-<!-- schema generated by tfplugindocs -->
-## Schema
-
-### Required
-
-- `name` (String) A unique name for the role
-
-### Optional
-
-- `description` (String) Description
-- `landing_url` (String) An optional override for the default landing page after login for a user.
-- `multitenant` (Boolean) Multitenant roles are copied to all tenant accounts and kept in sync until a sub-tenant user modifies their copy of the role. *Only available to master tenant*
-- `multitenant_locked` (Boolean) Multitenant Locked, prevents sub-tenant users from modifying their copy of multienant roles. *Only available to master tenant*
-- `permissions` (String) A JSON document containing the set of permissions to assign to the role
-- `role_type` (String) Role type
-
-### Read-Only
-
-- `id` (Number) The ID of the role
