+++ internal/subproviders/morpheus/datasources/rolepermissions/datasource_test.go
@@ -1,5 +1,7 @@
 // (C) Copyright 2025 Hewlett Packard Enterprise Development LP
 
+//go:build experimental
+
 package rolepermissions_test
 
 import (
