+++ internal/subproviders/morpheus/resources/role/morpheus_role.md.tmpl
